+++ packages/shakedown_core/lib/utils/pwa_detection.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:shakedown_core/utils/web_perf_hint.dart';
-import 'package:shakedown_core/utils/pwa_detection_stub.dart'
-    if (dart.library.js_interop) 'package:shakedown_core/utils/pwa_detection_web.dart'
+import 'pwa_detection_stub.dart'
+    if (dart.library.js_interop) 'pwa_detection_web.dart'
     as impl;
 
 enum WebRuntimeProfile {