+++ packages/shakedown_core/lib/utils/web_perf_hint_noop.dart
@@ -1,1 +1,3 @@
 bool isLikelyLowPowerWebDevice() => false;
+bool isSafariWeb() => false;
+bool isMobileWeb() => false;